moved SelfTestDesc to package pepper.core to avoid classloader problem
