+++ pepper-framework/src/main/java/org/corpus_tools/pepper/core/SelfTestDesc.java
@@ -1,4 +1,4 @@
-package org.corpus_tools.pepper.modules;
+package org.corpus_tools.pepper.core;
 
 import java.io.BufferedReader;
 import java.io.File;
