+++ pepper-framework/src/test/java/org/corpus_tools/pepper/modules/coreModules/tests/TextImporterTest.java
@@ -35,9 +35,9 @@ import org.corpus_tools.pepper.common.ModuleFitness;
 import org.corpus_tools.pepper.common.Pepper;
 import org.corpus_tools.pepper.common.ModuleFitness.FitnessFeature;
 import org.corpus_tools.pepper.core.ModuleFitnessChecker;
+import org.corpus_tools.pepper.core.SelfTestDesc;
 import org.corpus_tools.pepper.exceptions.PepperTestException;
 import org.corpus_tools.pepper.modules.PepperImporter;
-import org.corpus_tools.pepper.modules.SelfTestDesc;
 import org.corpus_tools.pepper.modules.coreModules.DoNothingImporter;
 import org.corpus_tools.pepper.modules.coreModules.TextImporter;
 import org.corpus_tools.pepper.modules.coreModules.TextImporter.TextMapper;
